Scaled Level Mockups
Placed environment boxes under static actors title Room A, B, and C Scaling Actor. Scaling the actor uniformly scales the objects underneath them.

We may still need to scale the environment more, but it is easy to do so now.

I moved the teleporters and turret, but I did not adjust where the teleporters send the player because I don't know how to.

David Trefry